test: 수익률 계산 테스트 코드 추가
